Split Hotspots spec and extract reusable test helpers
REDMINE-20853
