+++ entry_types/scrolled/package/spec/contentElements/hotspots/Hotspots/imageRendering-spec.js
@@ -1,0 +1,88 @@
+import React from 'react';
+
+import {Hotspots} from 'contentElements/hotspots/Hotspots';
+
+import {renderInContentElement} from 'pageflow-scrolled/testHelpers';
+import '@testing-library/jest-dom/extend-expect'
+
+import 'support/animateStub';
+import 'support/fakeResizeObserver';
+import 'support/scrollTimelineStub';
+
+describe('Hotspots', () => {
+  it('does not render images by default', () => {
+    const seed = {
+      imageFileUrlTemplates: {large: ':id_partition/image.webp'},
+      imageFiles: [{id: 1, permaId: 100}]
+    };
+    const configuration = {
+      image: 100
+    };
+
+    const {queryByRole} = renderInContentElement(
+      <Hotspots configuration={configuration} />, {seed}
+    );
+
+    expect(queryByRole('img')).toBeNull();
+  });
+
+  it('renders image when element should load', () => {
+    const seed = {
+      imageFileUrlTemplates: {
+        large: ':id_partition/large/image.webp',
+        ultra: ':id_partition/ultra/image.webp'
+      },
+      imageFiles: [{id: 1, permaId: 100}]
+    };
+    const configuration = {
+      image: 100
+    };
+
+    const {getByRole, simulateScrollPosition} = renderInContentElement(
+      <Hotspots configuration={configuration} />, {seed}
+    );
+    simulateScrollPosition('near viewport');
+
+    expect(getByRole('img')).toHaveAttribute('src', '000/000/001/large/image.webp');
+  });
+
+  it('supports portrait image', () => {
+    const seed = {
+      imageFileUrlTemplates: {large: ':id_partition/image.webp'},
+      imageFiles: [{id: 1, permaId: 100}, {id: 2, permaId: 101}]
+    };
+    const configuration = {
+      image: 100,
+      portraitImage: 101
+    };
+
+    window.matchMedia.mockPortrait();
+    const {getByRole, simulateScrollPosition} = renderInContentElement(
+      <Hotspots configuration={configuration} />, {seed}
+    );
+    simulateScrollPosition('near viewport');
+
+    expect(getByRole('img')).toHaveAttribute('src', '000/000/002/image.webp');
+  });
+
+  it('renders ultra image when pan zoom is enabled and element should load', () => {
+    const seed = {
+      imageFileUrlTemplates: {
+        large: ':id_partition/large/image.webp',
+        ultra: ':id_partition/ultra/image.webp'
+      },
+      imageFiles: [{id: 1, permaId: 100}]
+    };
+    const configuration = {
+      image: 100,
+      enablePanZoom: 'always'
+    };
+
+    const {getByRole, simulateScrollPosition} = renderInContentElement(
+      <Hotspots configuration={configuration} />, {seed}
+    );
+    simulateScrollPosition('near viewport');
+
+    expect(getByRole('img')).toHaveAttribute('src', '000/000/001/ultra/image.webp');
+  });
+});
