+++ entry_types/scrolled/package/spec/support/animateStub.js
@@ -1,0 +1,11 @@
+export const animateStub = {};
+
+beforeEach(() => {
+  animateStub.current = jest.fn(() => {
+    return {
+      cancel() {}
+    }
+  });
+
+  HTMLDivElement.prototype.animate = animateStub.current;
+});
