+++ entry_types/scrolled/package/spec/support/fakeIntersectionObserver.js
@@ -1,30 +1,57 @@
-const observedElements = new Set();
+import {act} from '@testing-library/react';
 
-global.IntersectionObserver = function(callback) {
-  return {
-    observe(el) {
-      observedElements.add(el);
-
-      const previousInvokeIntersectionCallback =
-        el.invokeIntersectionCallback;
-
-      el.invokeIntersectionCallback = (isIntersecting) => {
-        callback([{isIntersecting}]);
-
-        if (previousInvokeIntersectionCallback) {
-          previousInvokeIntersectionCallback(isIntersecting);
-        }
-      };
-    },
-
-    unobserve(el) {
-      observedElements.delete(el);
-    }
-  };
+export const fakeIntersectionObserver = {
+  byRoot(root) {
+    return this.instances.find(intersectionObserver =>
+      intersectionObserver.root === root
+    );
+  }
 };
 
+beforeEach(() => {
+  fakeIntersectionObserver.instances = new Set();
+  fakeIntersectionObserver.observedElements = new Set();
+});
+
+global.IntersectionObserver = function(callback, {threshold = 0, root} = {}) {
+  if (root && fakeIntersectionObserver.byRoot(root)) {
+    throw new Error('Did not except more than one intersection observer per root');
+  }
+
+  this.root = root;
+
+  this.observe = function(target) {
+    fakeIntersectionObserver.observedElements.add(target);
+
+    const previousInvokeIntersectionCallback =
+      target.invokeIntersectionCallback;
+
+    target.invokeIntersectionCallback = (isIntersecting) => {
+      callback([{target, isIntersecting, intersectionRatio: threshold}]);
+
+      if (previousInvokeIntersectionCallback) {
+        previousInvokeIntersectionCallback(isIntersecting);
+      }
+    };
+  };
+
+  this.unobserve = function(el) {
+    fakeIntersectionObserver.observedElements.delete(el);
+  };
+
+  this.disconnect = function() {};
+};
+
+export function simulateIntersecting(target) {
+  if (!target.invokeIntersectionCallback) {
+    throw new Error(`Intersection observer does not currently observe ${target}.`);
+  }
+
+  act(() => target.invokeIntersectionCallback(true));
+}
+
 export function simulateScrollingIntoView(visibleEl) {
-  observedElements.forEach(el => {
+  fakeIntersectionObserver.observedElements.forEach(el => {
     if (visibleEl.contains(el) || el.contains(visibleEl)) {
       el.invokeIntersectionCallback(true);
     }
@@ -32,7 +59,7 @@ export function simulateScrollingIntoView(visibleEl) {
 }
 
 export function simulateScrollingOutOfView(hiddenEl) {
-  observedElements.forEach(el => {
+  fakeIntersectionObserver.observedElements.forEach(el => {
     if (hiddenEl.contains(el) || el.contains(hiddenEl)) {
       el.invokeIntersectionCallback(false);
     }
