+++ entry_types/scrolled/package/spec/support/scrollTimelineStub.js
@@ -1,0 +1,10 @@
+export const scrollTimelineStub = {};
+
+beforeEach(() => {
+  scrollTimelineStub.instances = [];
+
+  window.ScrollTimeline = function(options) {
+    this.options = options;
+    scrollTimelineStub.instances.push(this);
+  };
+});
